added and id to html and a new variable to target this id in JS
Now we want to target the name which we know is the h1 in js;  so we go to the html and ad an id so we can then target it the JS.

we call the id"city"

in JS we call the city element( OR city variable) is the result of the city element; we use the # sign because we target an id. if we would target a class we would use a dot.