add an id in html to target the hardcoded temperature(12); add a new Variable in JS
go to html line of code 40

this code  line of code
          <div class="weather-app-temperature">12</div>
now becomes this

        <div class="weather-app-temperature" id="temperature">12</

Then go to the JS

the element with this id of temperature is going to be replaced by temperature. We create a new variable in line 2 of JS code.

so now this  code

console.log(response.data.temperature.current);

 is replaced by this code

let temperatureElement = document.querySelector("#temperature")
  temperatureElement.innerHTML = response.data.temperature.current;